Fix failing send when socket not ready
Fixes issue #207
When another message should be sent on a socket due to a pending reply,
which was not read yet, the message to be sent is put back in the queue.
As the read is already performed before that, nothing triggers a new flush_writes.
This makes the flushReads call asynchronous.