feat: use build manifest to load compatible binaries
This adds a `manifest.json` for the built binaries under built that allows the package to load the correct addon path during runtime. This PR also adds the libc/abi as information in the key, so it will avoid the conflicts between Musl/Glib, and loading will be faster since the compatible binaries are found deterministically.